split action and thunks in separate files and updated imports on all components
