+++ js/GameLostScreen.js
@@ -6,7 +6,7 @@ import {
   TouchableHighlight,
   Image
 } from "react-native";
-import { changeToUnset, changeToSpace } from "./store/gameActions";
+import { changeToUnset, changeToSpace } from "./store/thunks";
 import { connect } from "react-redux";
 
 class GameLostScreen extends Component {
