+++ js/Space.js
@@ -6,9 +6,8 @@ import {
   ViroSpinner
 } from "react-viro";
 import SingleObj from "./SingleObj";
-import allObjects from "./objects";
 import { connect } from "react-redux";
-import { displayAll } from "./store/gameActions";
+import { displayAll } from "./store/thunks";
 
 export default class Space extends Component {
   constructor() {
